Polaris table service api implementations and testing metadata json resource load (#488)

+++ metacat-connector-polaris/src/main/java/com/netflix/metacat/connector/polaris/PolarisConnectorTableService.java
@@ -9,6 +9,8 @@ import com.netflix.metacat.common.server.connectors.ConnectorRequestContext;
 import com.netflix.metacat.common.server.connectors.ConnectorTableService;
 import com.netflix.metacat.common.server.connectors.ConnectorUtils;
 import com.netflix.metacat.common.server.connectors.exception.ConnectorException;
+import com.netflix.metacat.common.server.connectors.exception.InvalidMetaException;
+import com.netflix.metacat.common.server.connectors.exception.TableAlreadyExistsException;
 import com.netflix.metacat.common.server.connectors.exception.TableNotFoundException;
 import com.netflix.metacat.common.server.connectors.model.TableInfo;
 import com.netflix.metacat.connector.hive.converters.HiveConnectorInfoConverter;
@@ -19,11 +21,12 @@ import com.netflix.metacat.connector.polaris.mappers.PolarisTableMapper;
 import com.netflix.metacat.connector.polaris.store.PolarisStoreConnector;
 import com.netflix.metacat.connector.polaris.store.entities.PolarisTableEntity;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataIntegrityViolationException;
 
 import javax.annotation.Nullable;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * table service for polaris connector.
@@ -69,11 +72,48 @@ public class PolarisConnectorTableService implements ConnectorTableService {
     @Override
     public void create(final ConnectorRequestContext requestContext, final TableInfo tableInfo) {
         final QualifiedName name = tableInfo.getName();
+        // check exists then create in non-transactional optimistic manner
+        if (exists(requestContext, name)) {
+            throw new TableAlreadyExistsException(name);
+        }
         try {
-            polarisConnector.createTable(name.getDatabaseName(), name.getTableName());
+            final PolarisTableMapper mapper = new PolarisTableMapper(name.getCatalogName());
+            final PolarisTableEntity entity = mapper.toEntity(tableInfo);
+            polarisConnector.createTable(entity.getDbName(), entity.getTblName(), entity.getMetadataLocation());
+        } catch (DataIntegrityViolationException | InvalidMetaException exception) {
+            throw new InvalidMetaException(name, exception);
         } catch (Exception exception) {
-            // TODO: distinguish between different constraint exceptions once data model supports get APIs
             final String msg = String.format("Failed creating polaris table %s", name);
+            log.error(msg, exception);
+            throw new ConnectorException(msg, exception);
+        }
+    }
+
+    /**
+     * {@inheritDoc}.
+     */
+    @Override
+    public void rename(
+        final ConnectorRequestContext context,
+        final QualifiedName oldName,
+        final QualifiedName newName
+    ) {
+        // check exists then rename in non-transactional optimistic manner
+        if (exists(context, newName)) {
+            throw new TableAlreadyExistsException(newName);
+        }
+        try {
+            final PolarisTableEntity table = polarisConnector
+                .getTable(oldName.getDatabaseName(), oldName.getTableName())
+                .orElseThrow(() -> new TableNotFoundException(oldName));
+            polarisConnector.saveTable(table.toBuilder().tblName(newName.getTableName()).build());
+        } catch (TableNotFoundException exception) {
+            log.error(String.format("Not found exception for polaris table %s", oldName), exception);
+            throw exception;
+        } catch (DataIntegrityViolationException exception) {
+            throw new InvalidMetaException(oldName, exception);
+        } catch (Exception exception) {
+            final String msg = String.format("Failed renaming polaris table %s", oldName);
             log.error(msg, exception);
             throw new ConnectorException(msg, exception);
         }
@@ -85,13 +125,11 @@ public class PolarisConnectorTableService implements ConnectorTableService {
     @Override
     public TableInfo get(final ConnectorRequestContext requestContext, final QualifiedName name) {
         try {
-            final Optional<PolarisTableEntity> polarisTableEntity =
-                polarisConnector.getTable(name.getDatabaseName(), name.getTableName());
-            if (!polarisTableEntity.isPresent()) {
-                throw new TableNotFoundException(name);
-            }
+            final PolarisTableEntity polarisTableEntity = polarisConnector
+                .getTable(name.getDatabaseName(), name.getTableName())
+                .orElseThrow(() -> new TableNotFoundException(name));
             final PolarisTableMapper mapper = new PolarisTableMapper(name.getCatalogName());
-            final TableInfo info = mapper.toInfo(polarisTableEntity.get());
+            final TableInfo info = mapper.toInfo(polarisTableEntity);
             final String tableLoc = HiveTableUtil.getIcebergTableMetadataLocation(info);
             final IcebergTableWrapper icebergTable = icebergTableHandler.getIcebergTable(
                 name, tableLoc, connectorContext.getConfig().isIcebergCacheEnabled());
@@ -153,16 +191,64 @@ public class PolarisConnectorTableService implements ConnectorTableService {
         }
     }
 
+
+    /**
+     * {@inheritDoc}.
+     */
+    @Override
+    public boolean exists(final ConnectorRequestContext requestContext, final QualifiedName name) {
+        try {
+            return polarisConnector.tableExists(name.getDatabaseName(), name.getTableName());
+        } catch (Exception exception) {
+            final String msg = String.format("Failed exists polaris table %s", name);
+            log.error(msg, exception);
+            throw new ConnectorException(msg, exception);
+        }
+    }
+
     /**
      * {@inheritDoc}.
      */
     @Override
     public void delete(final ConnectorRequestContext requestContext, final QualifiedName name) {
+        // check exists then delete in non-transactional optimistic manner
+        if (!exists(requestContext, name)) {
+            throw new TableNotFoundException(name);
+        }
         try {
             polarisConnector.deleteTable(name.getDatabaseName(), name.getTableName());
+        } catch (DataIntegrityViolationException exception) {
+            throw new InvalidMetaException(name, exception);
         } catch (Exception exception) {
-            // TODO: handle db not exist once data model supports get APIs
             final String msg = String.format("Failed deleting polaris table %s", name);
+            log.error(msg, exception);
+            throw new ConnectorException(msg, exception);
+        }
+    }
+
+    /**
+     * {@inheritDoc}.
+     */
+    @Override
+    public List<TableInfo> list(
+        final ConnectorRequestContext requestContext,
+        final QualifiedName name,
+        @Nullable final QualifiedName prefix,
+        @Nullable final Sort sort,
+        @Nullable final Pageable pageable
+    ) {
+        try {
+            final PolarisTableMapper mapper = new PolarisTableMapper(name.getCatalogName());
+            final String tableFilter = (prefix != null && prefix.isTableDefinition()) ? prefix.getTableName() : "";
+            final List<PolarisTableEntity> tbls =
+                polarisConnector.getTableEntities(name.getDatabaseName(), tableFilter);
+            if (sort != null) {
+                ConnectorUtils.sort(tbls, sort, Comparator.comparing(t -> t.getTblName()));
+            }
+            return ConnectorUtils.paginate(tbls, pageable).stream()
+                .map(t -> mapper.toInfo(t)).collect(Collectors.toList());
+        } catch (Exception exception) {
+            final String msg = String.format("Failed polaris list tables %s using prefix %s", name, prefix);
             log.error(msg, exception);
             throw new ConnectorException(msg, exception);
         }
